Migration to generated Win32 bindings (#247)
* migrate to autogenerated Win32 bindings

- Delete old manual Win32Types.kt and Win32_h.kt
- Add win32_all_h.kt (generated via jextract from Windows headers)
- Add Win32Constants.kt with constants + helpers extracted from Win32_h.kt
- Replace MethodHandle invokeExact patterns with generated PascalCase functions
  across KadreWndProc, Win32EventLoop, Win32EventLoopProxy, Win32Window,
  DxgiOutputDuplicator, Win32ScreenCapturer, Win32ExtensionTypes, Win32Monitor
- Fix tests to match new API (type changes, no-null assertion removal)
- Add typealias re-exports in Win32Constants.kt for backward compat

* add --init-method flag for deferred Win32 binding init

- Update kextract submodule to feat/init-method branch (5f6d895)
- Pass --init-method to kextract in regen script
- CI will regenerate win32_all_h.kt with init() pattern

+++ kadre-win32/src/jvmTest/kotlin/org/graphiks/kadre/win32/SmokeTest.kt
@@ -36,10 +36,10 @@ class SmokeTest {
     fun `the Win32 type aliases compile and are assignable`() {
         // Verifies that the typealiases are usable as ordinary Kotlin types
         val dword: DWORD = 0xFFFF_FFFFL
-        val word: WORD = 0xFFFF
-        val uint: UINT = 0L
+        val word: WORD = 0xFFFF.toShort()
+        val uint: UINT = 0
         val ulongPtr: ULONG_PTR = 0L
-        val long: LONG = -1
+        val long: LONG = -1L
         val longPtr: LONG_PTR = -1L
         val lresult: LRESULT = 0L
         val wparam: WPARAM = 0L
@@ -50,7 +50,7 @@ class SmokeTest {
         val hmodule: HMODULE = 0L
         val hdc: HDC = 0L
         val hmenu: HMENU = 0L
-        val atom: ATOM = 0
+        val atom: ATOM = 0.toShort()
 
         assertNotNull(dword)
         assertNotNull(word)
